Added an example demoing how to use grpc with a "scheduler" thread.
I think that this is the way the backend is going to be implemented.
The data will consist of a gardening plan / schedule, with (?) options
to overwrite the current state of every valve.

The "scheduler" thread will wake up about once a second, check the data
to decide what the current state should be and set it to that.

The grpc server will simply update the data as requested, so that the
next time the "scheduler" wakes up, it will update its state.

I think this project is small enough for this kind of design, though I'm
pretty sure you're supposed to "share data by communcating", and not
"communicate by sharing data". This seems like the simplest option.